+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 4.4.0:
 * Added `grizzled.datetime.EnhancedDuration` class, with a `humanize` method
   that provides more human-readable string representations for `Duration`
   objects.
+* `grizzled.net.UDPDatagramSocket` now extends `AutoCloseable`.
 
 Version 4.3.0: